Simplify the starting project
