+++ gatsby-config.js
@@ -6,13 +6,6 @@ module.exports = {
   plugins: [
     'gatsby-plugin-react-helmet',
     {
-      resolve: 'gatsby-source-filesystem',
-      options: {
-        name: 'images',
-        path: `${__dirname}/src/images`,
-      },
-    },
-    {
       resolve: 'gatsby-plugin-prefetch-google-fonts',
       options: {
         fonts: [
@@ -32,12 +25,6 @@ module.exports = {
     {
       resolve: 'gatsby-plugin-manifest',
       options: {
-        name: 'gatsby-starter-default',
-        short_name: 'starter',
-        start_url: '/',
-        background_color: '#663399',
-        theme_color: '#663399',
-        display: 'minimal-ui',
         icon: 'src/images/favicon.png',
       },
     },
