+++ src/components/Layout.js
@@ -1,5 +1,4 @@
 import React from 'react'
-import PropTypes from 'prop-types'
 import Header from './Header'
 import Footer from './Footer'
 
@@ -10,14 +9,5 @@ const Layout = ({ isHomepage, children }) => (
     <Footer />
   </>
 )
-
-Layout.propTypes = {
-  isHomepage: PropTypes.bool,
-  children: PropTypes.node.isRequired,
-}
-
-Layout.defaultProps = {
-  isHomepage: false,
-}
 
 export default Layout
