--- src/components/Layout/package.json
@@ -1,3 +1,0 @@
-{
-  "main": "Layout.js"
-}
