+++ src/styles/style.css
@@ -93,7 +93,7 @@ footer {
 }
 .text-section-1col img,
 .text-section-2col img,
-.gallery-image {
+.gallery img {
   margin-bottom: 1rem;
 }
 .text-section-1col p:last-child,
